+++ src/java/com/threerings/bang/game/data/piece/HeroInfluence.java
@@ -66,13 +66,13 @@ public class HeroInfluence extends Influence
     @Override // documentation inherited
     public int adjustAttack (Piece target, int damage)
     {
-        return (int)(damage + 3 * _level);
+        return (damage + 3 * _level);
     }
 
     @Override // documentation inherited
     public int adjustDefend (Piece shooter, int damage)
     {
-        return Math.max(0, (int)(damage  - 3 * _level));
+        return Math.max(0, (damage  - 3 * _level));
     }
 
     @Override // documentation inherited
